Refactor summarize module to improve organization and documentation
